requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
faiss-cpu
openai
pandas
fastapi
uvicorn
langchain_openai
streamlit